Rearrange query.go and remove dead code
